Update settings for sbt-sonatype 3.0
