+++ .scalafmt.conf
@@ -1,3 +1,4 @@
-maxColumn = 180
+version = 2.0.1
+maxColumn = 120
 style = defaultWithAlign
 optIn.breaksInsideChains = true
