+++ project/build.properties
@@ -1,2 +1,1 @@
-sbt.version=1.2.7
-
+sbt.version=1.3.0-RC5
