+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "2.7-SNAPSHOT"
+version in ThisBuild := "3.0-SNAPSHOT"
